+++ Tests/RxSwift/mapToTests.swift
@@ -38,7 +38,7 @@ class MapToTests: XCTestCase {
     }
 
     func testReplaceWithResultValues() {
-        //test elements values and type
+        // test elements values and type
         let correctValues = [
             next(0, "candy"),
             next(0, "candy"),